Merge code branch dev to prod

See merge request chien.tranvan/vtes-backend!4

+++ src/main/java/com/vtes/repository/RefreshTokenRepository.java
@@ -13,7 +13,7 @@ import com.vtes.entity.User;
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
 	Optional<RefreshToken> findByToken(String token);
 
-	Optional<RefreshToken> findByUserId(Long userId);
+	Optional<RefreshToken> findByUserId(Integer userId);
 
 	@Modifying
 	int deleteByUser(User user);
